rename classicSpinnerConfig
