+++ src/components/spinner/classicSpinnerConfig.js
