+++ src/components/spinner/spinner.style.js
@@ -1,7 +1,7 @@
 import styled, { keyframes } from 'styled-components';
 import PropTypes from 'prop-types';
 import OptionsHelper from '../../utils/helpers/options-helper';
-import { classicSpinnerColors, classicSpinnerSizes } from './classicSpinnerValues';
+import { classicSpinnerColors, classicSpinnerSizes } from './classicSpinnerConfig';
 
 const spinnerAnimation = keyframes`
   from { transform: rotate(0deg); }
